+++ packages/gamut-labs/src/experimental/GlobalHeader/__tests__/GlobalHeader-test.tsx
@@ -39,14 +39,20 @@ const anonLandingHeaderProps: GlobalHeaderProps = {
 
 const anonLoginHeaderProps: GlobalHeaderProps = {
   action,
-  renderSearch: () => <IconButton icon={SearchIcon} />,
+  renderSearch: {
+    desktop: () => <IconButton icon={SearchIcon} />,
+    mobile: () => <IconButton icon={SearchIcon} />,
+  },
   type: 'anon',
   variant: 'login',
 };
 
 const anonSignUpHeaderProps: GlobalHeaderProps = {
   action,
-  renderSearch: () => <IconButton icon={SearchIcon} />,
+  renderSearch: {
+    desktop: () => <IconButton icon={SearchIcon} />,
+    mobile: () => <IconButton icon={SearchIcon} />,
+  },
   type: 'anon',
   variant: 'signup',
 };
@@ -65,8 +71,14 @@ const proHeaderProps: GlobalHeaderProps = {
 
 const renderElementProps: GlobalHeaderProps = {
   action,
-  renderSearch: () => <IconButton icon={SearchIcon} />,
-  renderNotifications: () => <IconButton icon={BellIcon} />,
+  renderSearch: {
+    desktop: () => <IconButton icon={SearchIcon} />,
+    mobile: () => <IconButton icon={SearchIcon} />,
+  },
+  renderNotifications: {
+    desktop: () => <IconButton icon={BellIcon} />,
+    mobile: () => <IconButton icon={BellIcon} />,
+  },
   type: 'pro',
   user,
 };
@@ -79,42 +91,70 @@ const renderGlobalHeader = (props: GlobalHeaderProps) => {
   );
 };
 
+jest.mock('../../../lib/breakpointHooks', () => ({
+  useBreakpointAtOrAbove: jest
+    .fn(() => true) // default implementation: desktop true
+    .mockImplementationOnce(() => {
+      return true;
+    })
+    .mockImplementationOnce(() => {
+      return false;
+    }),
+}));
+
 describe('GlobalHeader', () => {
+  describe('responsiveness', () => {
+    beforeEach(() => {
+      renderGlobalHeader(anonHeaderProps);
+    });
+
+    test('desktop', () => {
+      expect(screen.queryByTestId('header-mobile-menu')).not.toBeVisible();
+    });
+    test('mobile', () => {
+      expect(screen.getByTestId('header-mobile-menu')).toBeVisible();
+    });
+  });
+
   describe('anonymous users', () => {
     beforeEach(() => {
       renderGlobalHeader(anonHeaderProps);
     });
 
+    /* since we're using css to toggle the display of the header between desktop & mobile, these tests check for visibility of elements
+     & use 'getAllByTestId' b/c there will be duplicate elements in the DOM (since mobile & desktop render some of the same app header items) */
     test('logo', () => {
-      screen.getByTestId('header-logo');
+      const logoElements = screen.getAllByTestId('header-logo');
+      expect(logoElements[0]).toBeVisible(); // desktop
+      expect(logoElements[1]).not.toBeVisible(); // mobile
     });
 
     test('courseCatalog', () => {
-      screen.getByText(courseCatalog.text);
+      screen.getAllByText(courseCatalog.text);
     });
 
     test('resourcesDropdown', () => {
-      screen.getByText(resourcesDropdown.text);
+      screen.getAllByText(resourcesDropdown.text);
     });
 
     test('communityDropdown', () => {
-      screen.getByText(communityDropdown.text);
+      screen.getAllByText(communityDropdown.text);
     });
 
     test('plansPricingDropdown', () => {
-      screen.getByText(pricingDropdown.text);
+      screen.getAllByText(pricingDropdown.text);
     });
 
     test('forEnterprise', () => {
-      screen.getByText(forBusiness.text);
+      screen.getAllByText(forBusiness.text);
     });
 
     test('login', () => {
-      screen.getByText(login.text);
+      screen.getAllByText(login.text);
     });
 
     test('signup', () => {
-      screen.getByText(signUp.text);
+      screen.getAllByText(signUp.text);
     });
   });
 
@@ -129,7 +169,7 @@ describe('GlobalHeader', () => {
       });
 
       test('shows login', () => {
-        screen.getByText(login.text);
+        screen.getAllByText(login.text);
       });
 
       test('does not show signup', () => {
@@ -151,7 +191,7 @@ describe('GlobalHeader', () => {
       });
 
       test('shows signup', () => {
-        screen.getByText(signUp.text);
+        screen.getAllByText(signUp.text);
       });
     });
 
@@ -161,11 +201,11 @@ describe('GlobalHeader', () => {
       });
 
       test('shows search', () => {
-        screen.getByTitle('Search Icon');
+        screen.getAllByTitle('Search Icon');
       });
 
       test('shows login', () => {
-        screen.getByText(login.text);
+        screen.getAllByText(login.text);
       });
 
       test('does not show sign up', () => {
@@ -180,7 +220,7 @@ describe('GlobalHeader', () => {
     });
 
     test('logo', () => {
-      screen.getByTestId('header-logo');
+      screen.getAllByTestId('header-logo');
     });
 
     test('myHome', () => {
@@ -222,7 +262,7 @@ describe('GlobalHeader', () => {
     });
 
     test('proLogo', () => {
-      screen.getByTestId('header-pro-logo');
+      screen.getAllByTestId('header-pro-logo');
     });
 
     test('myHome', () => {
@@ -256,7 +296,7 @@ describe('GlobalHeader', () => {
     });
 
     test('notifications', () => {
-      screen.getByTitle('Bell Icon');
+      screen.getAllByTitle('Bell Icon');
     });
   });